requirements.txt
python-telegram-bot==20.1
google-api-python-client==2.82.0
google-auth-httplib2==0.1.0
loguru==0.6.0
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
asyncpg==0.27.0 "no para mysql"
pymysql==1.0.2